Redirect flow instead of Javascript flow
Instead of using Javascript in the browser that pop-ups for
Google authentication and then pushes an ID token to the IdP
the browser is now redirected to Google for authentication and
then Google redirects back to the IdP with a one-time
authorization code. The IdP then exchanges the authorization
code for the ID token.